KOGTIO-5804: prefix package as org.kie.kogito - pmml-[quarkus|springboot]-example
