+++ pmml-quarkus-example/src/test/java/org/kie/kogito/pmml/quarkus/example/LinRegTest.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.pmml.quarkus.example;
+
+import java.util.Collections;
+import java.util.Map;
+
+import org.junit.jupiter.api.Test;
+
+import io.quarkus.test.junit.QuarkusTest;
+
+import static org.kie.kogito.pmml.quarkus.example.CommonTestUtils.testDescriptive;
+import static org.kie.kogito.pmml.quarkus.example.CommonTestUtils.testResult;
+
+@QuarkusTest
+class LinRegTest {
+
+    private static final String BASE_PATH = "/LinReg";
+    private static final String TARGET = "fld4";
+
+    @Test
+    void testEvaluateLinRegResult() {
+        String inputData = "{\"fld1\":3.0, \"fld2\":2.0, \"fld3\":\"y\"}";
+        testResult(inputData, BASE_PATH, TARGET, 52.5f);
+    }
+
+    @Test
+    void testEvaluateLinRegDescriptive() {
+        String inputData = "{\"fld1\":3.0, \"fld2\":2.0, \"fld3\":\"y\"}";
+        final Map<String, Object> expectedResultMap = Collections.singletonMap(TARGET, 52.5f);
+        testDescriptive(inputData, BASE_PATH, TARGET, expectedResultMap);
+    }
+
+}
