+++ pmml-quarkus-example/src/test/java/org/kie/kogito/pmml/quarkus/example/NativeLinRegTestIT.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.pmml.kogito.quarkus.example;
+package org.kie.kogito.pmml.quarkus.example;
 
 import io.quarkus.test.junit.NativeImageTest;
 
 @NativeImageTest
-public class NativeSampleMineTestIT extends SampleMineTest {
+public class NativeLinRegTestIT extends LinRegTest {
 
     // Execute the same tests but in native mode.
 }
