+++ pmml-quarkus-example/src/test/java/org/kie/kogito/pmml/quarkus/example/NativeMiningModelTestIT.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.pmml.kogito.quarkus.example;
+package org.kie.kogito.pmml.quarkus.example;
 
 import io.quarkus.test.junit.NativeImageTest;
 
 @NativeImageTest
-public class NativeLinRegTestIT extends LinRegTest {
+public class NativeMiningModelTestIT extends MiningModelTest {
 
     // Execute the same tests but in native mode.
 }
