+++ pmml-quarkus-example/src/test/java/org/kie/kogito/pmml/quarkus/example/NativeSimpleScorecardTestIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.pmml.kogito.quarkus.example;
+package org.kie.kogito.pmml.quarkus.example;
 
 import io.quarkus.test.junit.NativeImageTest;
