+++ pmml-springboot-example/src/test/java/org/kie/kogito/pmml/springboot/example/LinRegTest.java
@@ -13,23 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.pmml.kogito.quarkus.example;
+package org.kie.kogito.pmml.springboot.example;
 
 import java.util.Collections;
 import java.util.Map;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.annotation.DirtiesContext;
 
-import io.quarkus.test.junit.QuarkusTest;
+import io.restassured.RestAssured;
 
-import static org.kie.pmml.kogito.quarkus.example.CommonTestUtils.testDescriptive;
-import static org.kie.pmml.kogito.quarkus.example.CommonTestUtils.testResult;
+import static org.kie.kogito.pmml.springboot.example.CommonTestUtils.testDescriptive;
+import static org.kie.kogito.pmml.springboot.example.CommonTestUtils.testResult;
 
-@QuarkusTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = KogitoSpringbootApplication.class)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 class LinRegTest {
 
     private static final String BASE_PATH = "/LinReg";
     private static final String TARGET = "fld4";
+
+    @LocalServerPort
+    private int port;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.port = port;
+    }
 
     @Test
     void testEvaluateLinRegResult() {
